Removed command satisfy option
The command satisfy option was allowing a
false positive on python installation for
the virtualenv.  The package satisfiers
all require python development packages,
but the command option only checks for
python itself.  The development packages
for openSUSE at least, include a python-config
utility that isn't included in the regular
python package.  Something like python-config
should be checked, but even that isn't enough.
The real check should be if Python.h is in
the virtualenv include directories.  What's
the best way to do that?