gr-osmosdr: add missing dependencies
The gr-osmosdr recipe is missing some of the optional dependencies 
needed to build modules for its supported SDR boards. As a result, 
support for Airspy HF and FUNcube Dongle hardware is not built. This 
change adds the missing dependencies.

In addition, SDRplay support is not built, even if the required library 
is present, because gr-osmosdr requires the ENABLE_NONFREE config option 
to be set. I've added it to config_opt.